Remove limitations on finding choices w/ conditionals
Previously you could not have conditional branches directly after a speaker line where one branch led to choices and another didn't. That restriction is now gone. While we try to precompute the possibility of choices appearing, we don't require that to be 100% accurate and conditionals will be evaluated when probing for choices as well. Similarly, choices which only appear via gosubs are now much more robustly discovered.