Use `ETextIdenticalModeFlags::DeepCompare` flag for text `IdenticalTo` comparsion
With the default flag `ETextIdenticalModeFlags::None`, the first step of the comparison checks if the `TextData` pointers of the localized string pointers are identical.
In the test case, the `TextData` pointers differ as the `FSUDSScriptImporter` creates them in separate `FTextInspector::GetTextId` calls.
In the second step, pointers to the localized strings are compared.
The localized string pointers can be `nullptr` if we don't have localization data.
In such a case, the comparison returns false, and the tests fail.

With the flag `ETextIdenticalModeFlags::DeepCompare`, yet another step is executed.
It checks if we create the test the same way (check `FTextHistory_StringTableEntry::FStringTableReferenceData::IsIdentical`).