Force "en-US" culture in some tests
It fixes false-positive failures in tests with number, gender, and plural form formatting.

Number and plural formatting are locale-specific, so we must set it to a predefined value to produce the expected output in tests.

E.g. In Polish `,` is used as the decimal separator instead of `.`.
Declaring people and numbers is much more complicated than in English.